move the VRT stack test to `test_readers.py`
